Move hw->sw frame mapping into a separate class
